chore: add reset example
Make procedural data provider available outside of doc.
