+++ doc/src/theme/ReactLiveScope/index.js
@@ -18,9 +18,11 @@ import {ViewerComponent} from '../../js/components/ViewerComponent';
 import * as animation from '../../js/examples/extend-animation';
 import * as flycontrols from '../../js/examples/extend-fly-controls';
 import * as graphics from '../../js/examples/extend-graphics-developer';
-import * as procedural from '../../js/examples/extend-procedural-data-provider';
 import * as threerenderer from '../../js/examples/extend-three-renderer';
 import * as webglrenderer from '../../js/examples/extend-webgl-renderer';
+
+import * as procedural from '../../js/utils/ProceduralDataProvider';
+import * as image from '../../js/utils/image';
 
 import '../../mapillary-js/dist/mapillary.css';
 
@@ -31,9 +33,10 @@ const ReactLiveScope = {
   animation,
   flycontrols,
   graphics,
-  procedural,
   webglrenderer,
   threerenderer,
+  procedural,
+  image,
   ...mapillary,
   three,
   ViewerComponent,
